+++ src/lib/ngx-fileupload/components/ngx-fileupload-item.component.ts
@@ -66,8 +66,10 @@ export class NgxFileUploadItemComponent implements OnInit {
      */
     @ViewChild("defaultUploadItem", {static: true})
     @Input()
-    public set template(tpl: TemplateRef<any>) {
-        this.itemTpl = tpl || this.itemTpl;
+    public set template(tpl: TemplateRef<FileUploadItemContext>) {
+        if (tpl !== undefined) {
+            this.itemTpl = tpl;
+        }
     }
 
     /**